% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constructors.R
\name{breeding_scheme}
\alias{breeding_scheme}
\title{Define a minimal breeding scheme}
\usage{
breeding_scheme(
  generations,
  population_size,
  selected_parents,
  offspring_per_mating
)
}
\arguments{
\item{generations}{Number of offspring generations to simulate.}

\item{population_size}{Number of individuals in each generation.}

\item{selected_parents}{Number of parents randomly selected each generation.}

\item{offspring_per_mating}{Number of offspring produced by each mating.}
}
\value{
An object of class \code{qgsim_scheme}.
}
\description{
Define a minimal breeding scheme
}
