% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plan.R
\name{qgsim_compile}
\alias{qgsim_compile}
\title{Compile a breeding-scheme specification}
\usage{
qgsim_compile(scheme, seed = 1L, outputs = output_spec())
}
\arguments{
\item{scheme}{A \code{qgsim_scheme} created by \code{\link[=breeding_scheme]{breeding_scheme()}}.}

\item{seed}{Integer random seed.}

\item{outputs}{A \code{qgsim_outputs} created by \code{\link[=output_spec]{output_spec()}}.}
}
\value{
A validated object of class \code{qgsim_plan}.
}
\description{
Compile a breeding-scheme specification
}
