% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/backend.R
\name{qgsim_list_backends}
\alias{qgsim_list_backends}
\title{List registered qgsim backends}
\usage{
qgsim_list_backends()
}
\value{
Character vector of backend names.
}
\description{
List registered qgsim backends
}
