% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{qgsim_metrics}
\alias{qgsim_metrics}
\title{Extract standardized simulation metrics}
\usage{
qgsim_metrics(result)
}
\arguments{
\item{result}{A \code{qgsim_result}.}
}
\value{
A data frame.
}
\description{
Extract standardized simulation metrics
}
