% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/result.R
\name{qgsim_populations}
\alias{qgsim_populations}
\title{Extract standardized simulated populations}
\usage{
qgsim_populations(result)
}
\arguments{
\item{result}{A \code{qgsim_result}.}
}
\value{
A data frame.
}
\description{
Extract standardized simulated populations
}
