% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsim_repl.R
\name{qgsim_repl}
\alias{qgsim_repl}
\title{Simulate evolution (with replicates and summaries)}
\usage{
qgsim_repl(
  nreps = 1,
  summaries = c(),
  npops = 2,
  mig = 0,
  Ne = 500,
  theta0 = 0,
  z0 = 0,
  h2 = 0.5,
  Gcor = 0.2,
  omega11 = 1,
  omega22 = 1,
  omegaCor = 0,
  model = "Brownian",
  ngens = 100,
  tsd = 0.02,
  tmn = 0.01
)
}
\arguments{
\item{nreps}{number of replicates to run}

\item{summaries}{a vector of strings containing any summaries to be included in the results}

\item{npops}{number of populations to simulate.}

\item{mig}{the migration rate, this is the total proportion of individuals in each population made up of migrants from the other populations.}

\item{theta0}{this specifies the initial optimal value for the traits, you can supply a single value (applied to both traits), one value per trait, or a matrix with one value per population (row) and trait (column).}

\item{z0}{this specifies the initial trait values, you can supply a single value (applied to both traits), one value per trait, or a matrix with one value per population (row) and trait (column).}

\item{h2}{the trait heritabilities, assumed to be the same for both traits (must be between 0 and 1).}

\item{Gcor}{the genetic correlation between the pair of traits (must be between -1 and 1).}

\item{omega11}{denotes the intensity of selection (curvature of the adaptive landscape) with respect to trait 1, larger values result in weaker selection \link{default = 1}.}

\item{omega22}{denotes the intensity of selection (curvature of the adaptive landscape) with respect to trait 2, larger values result in weaker selection \link{default = 1}.}

\item{omegaCor}{denotes the strength of correlational selection, larger values denote stronger selection for combinations of trait 1 and trait 2 (must be between -1 and 1, 0 denotes independent selection on each trait) \link{default = 0}.}

\item{model}{denotes the model for adaptive peak movement, must be one of the following: "Browning", "Uncorrelated", or "Trend"; detailed descriptions are provided below.}

\item{ngens}{number of generations to simulate \link{default = 100}.}

\item{tsd}{standard deviation for peak movement, larger values denote larger (random) jumps in adaptive peak locations (must be a positive number).}

\item{tmn}{average direction shift in the location of the adaptive landscape, only relevant for the "Trend" model (can be negative or positive, but must be a single value for both traits).}
}
\value{
A list with two objects, the mean trait values (z) and adaptive peak locations (theta). Each of these is itself a list with one matrix per population. The population matrixes have two columns, one per trait, and one row per generation.
}
\description{
This function is a wrapper for the gqsim function which lets you run replicates and get summaries.
The summaries currently available include: mean rate of evolution ('mean_evol_rate', average change in each trait across populations and generations), evolutionary lag ('mean_evol_lag', average distance in bivaraiate trait space between the optimal phenotypes and z), correlations in evolution between traits ('mean_traitwise_corr', averaged over time and populations) and correlations in evolution across populations ('mean_popwise_corr', averaged over time and traits).
}
\details{
With the Brownian motion peak shift model, a random normal deviation with standard deviation tsd is added to the peak location each generation. With the Uncorrelated peak shift model, the peak location is follows a normal distribution with independent peak locations each generation. The peak location for each trait is specifically drawn from a normal distribution with mean 0 and standard deviation tst. Finally for Trend, the peak location shifts by a random normal deviation each generation with a mean tmn and standard deviation tsd. In all cases, peak shifts occur independently in each population.
}
