% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cvmatvis.R
\name{var_covar_mat_vis}
\alias{var_covar_mat_vis}
\title{Visualize variance/covariance matrices}
\usage{
var_covar_mat_vis(h2 = 0.5, Gcor = 0.2, omega11 = 1, omega22 = 1, omegaCor = 0)
}
\arguments{
\item{h2}{the trait heritabilities, assumed to be the same for both traits (must be between 0 and 1).}

\item{Gcor}{the genetic correlation between the pair of traits (must be between -1 and 1).}

\item{omega11}{denotes the intensity of selection (curvature of the adaptive landscape) with respect to trait 1, larger values result in weaker selection \link{default = 1}.}

\item{omega22}{denotes the intensity of selection (curvature of the adaptive landscape) with respect to trait 2, larger values result in weaker selection \link{default = 1}.}

\item{omegaCor}{denotes the strength of correlational selection, larger values denote stronger selection for combinations of trait 1 and trait 2 (must be between -1 and 1, 0 denotes independent selection on each trait) \link{default = 0}.}
}
\value{
The plot showing 95\% CI for both generated matrices
}
\description{
Visualize the G and omega variance/covariance matrices generated in qgsim
}
