% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsim-package.R
\docType{package}
\name{qgsim-package}
\alias{qgsim}
\alias{qgsim-package}
\title{qgsim: Common Interfaces for Breeding-Scheme Simulation}
\description{
qgsim separates breeding-scheme specification from simulation execution.
It compiles user inputs into a backend-neutral plan, dispatches that plan to
a registered backend, and returns a standardized result.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://psoerensen.github.io/qgsim/}
  \item \url{https://github.com/psoerensen/qgsim}
  \item Report bugs at \url{https://github.com/psoerensen/qgsim/issues}
}

}
\author{
\strong{Maintainer}: qgsim authors \email{qgsim@example.org}

}
\keyword{internal}
