% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/qgsim-package.R
\docType{package}
\name{qgsim-package}
\alias{qgsim-package}
\alias{qgsim}
\title{The 'qgsim' package.}
\description{
This package was developed for educational purposes. It can be used to simulate quantitative trait evolution on dynamic adapative landscapes for multiple populations. Two traits are modeled and these evolve by selection, drift and gene flow.
}
