\name{qgsim-package}
\alias{qgsim-package}
\alias{qgsim}
\docType{package}
\title{
\packageTitle{qgsim}
}
\description{
\packageDescription{qgsim}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{qgsim}
\packageIndices{qgsim}
~~ An overview of how to use the package, including the most important ~~
~~ functions ~~
}
\author{
\packageAuthor{qgsim}

Maintainer: \packageMaintainer{qgsim}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in ~~
~~ the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
